+++ sdks/go/pkg/beam/runners/prism/internal/environments.go
@@ -22,6 +22,8 @@ import (
 	"io"
 	"log/slog"
 	"os"
+	"os/exec"
+	"time"
 
 	fnpb "github.com/apache/beam/sdks/v2/go/pkg/beam/model/fnexecution_v1"
 	pipepb "github.com/apache/beam/sdks/v2/go/pkg/beam/model/pipeline_v1"
@@ -66,6 +68,16 @@ func runEnvironment(ctx context.Context, j *jobservices.Job, env string, wk *wor
 			logger.Error("unmarshing docker environment payload", "error", err)
 		}
 		return dockerEnvironment(ctx, logger, dp, wk, j.ArtifactEndpoint())
+	case urns.EnvProcess:
+		pp := &pipepb.ProcessPayload{}
+		if err := (proto.UnmarshalOptions{}).Unmarshal(e.GetPayload(), pp); err != nil {
+			logger.Error("unmarshing docker environment payload", "error", err)
+		}
+		go func() {
+			processEnvironment(ctx, pp, wk)
+			logger.Debug("environment stopped", slog.String("job", j.String()))
+		}()
+		return nil
 	default:
 		return fmt.Errorf("environment %v with urn %v unimplemented", env, e.GetUrn())
 	}
@@ -230,4 +242,23 @@ func dockerEnvironment(ctx context.Context, logger *slog.Logger, dp *pipepb.Dock
 	}()
 
 	return nil
+}
+
+func processEnvironment(ctx context.Context, pp *pipepb.ProcessPayload, wk *worker.W) {
+	cmd := exec.CommandContext(ctx, pp.GetCommand(), "--id="+wk.ID, "--provision_endpoint="+wk.Endpoint())
+
+	cmd.WaitDelay = time.Millisecond * 100
+	cmd.Stderr = os.Stderr
+	cmd.Stdout = os.Stdout
+	cmd.Env = os.Environ()
+
+	for k, v := range pp.GetEnv() {
+		cmd.Env = append(cmd.Environ(), fmt.Sprintf("%v=%v", k, v))
+	}
+	if err := cmd.Start(); err != nil {
+		return
+	}
+	// Job processing happens here, but orchestrated by other goroutines
+	// This call blocks until the context is cancelled, or the command exits.
+	cmd.Wait()
 }